Fix missing underscore in Yul function names

+++ test/libsolidity/semanticTests/inheritance/transient_storage_state_variable.sol
@@ -1,13 +1,13 @@
 contract A {
-    uint transient x;
-    int y;
+    uint24 transient x;
+    int24 y;
 }
 
 contract C is A {
-    uint w;
-    int transient z;
+    uint24 w;
+    int24 transient z;
 
-    function f() public returns (uint, int, uint, int) {
+    function f() public returns (uint24, int24, uint24, int24) {
         x += 1;
         y += 2;
         w += 3;
